Moved error logging to log.py from database.py
