+++ modules/database.py
@@ -6,51 +6,14 @@ import pymysql
 import os
 import time
 
+from modules.log import logging
+
 class database(object):
 
 	def __init__(self):
 
 		self.services = []
-
-
-	def log_error_locally(self, error_message):
-		if(os.path.exists("errorlog.txt") == True):
-			with open("errorlog.txt", "a+") as f:
-				error_message = str(time.strftime("%b %d %Y %H:%M:%S", time.gmtime(time.time()))) + "	" + str(error_message) + "\n"
-				f.write(error_message)
-				f.close()
-		else:
-			try:
-				f = open("errorlog.txt", "w+")
-				f.close()
-				self.log_error_locally(error_message)
-			except Exception as e:
-				print("Could not log error locally!")
-				print(e)
-
-	def log_error(self, error_message):
-		print(error_message)	
-		temp_connection = self.connection
-
-		with temp_connection.cursor() as cursor:
-			try:
-				#log the error locally
-				self.log_error_locally(error_message)
-
-				#log the error in the database
-				cursor.execute("""INSERT INTO exceptions (exception) VALUES ("%s")""" % (error_message))
-				temp_connection.commit()
-				print("succesfully logged exception in database")
-
-			except Exception as e:
-				#log error message locally
-				self.log_error_locally(error_message)
-
-				#log why couldn't write to database locally
-				self.log_error_locally(e)
-
-				print("!! Could not log exception!!")	
-
+		self.logging = logging()
 
 	def create_connection(self, db_config):
 		try:
@@ -65,7 +28,7 @@ class database(object):
 
 		except Exception as e:
 			print("COULD NOT ESTABLISH CONNECTION TO DATABASE. PLEASE CHECK config.py !STOPPING ...")
-			self.log_error(e)
+			self.logging.log_error(e)
 			exit()
 
 	def create_exceptions_table(self, connection):
@@ -80,11 +43,11 @@ class database(object):
 					print("Found 'exceptions' table!")
 				else:
 					print("COULD NOT ESTABLISH CONNECTION TO 'EXCEPTIONS' DATABSE! REASON:")
-					self.log_error(e)
+					self.logging.log_error(e)
 
 			except Exception as e:
 				print("COULD NOT ESTABLISH CONNECTION TO 'EXCEPTIONS' DATABSE! REASON:")
-				self.log_error(e)	
+				self.logging.log_error(e)	
 
 	def create_services_table(self, connection):
 		with connection.cursor() as cursor:
@@ -97,10 +60,10 @@ class database(object):
 				if("1050" in str(e)): #1050 is warning code for table allready exists
 					print("Found 'services' table!")
 				else:
-					self.log_error(e)
+					self.logging.log_error(e)
 
 			except Exception as e:
-				self.log_error(e)
+				self.logging.log_error(e)
 
 	def fetch_services(self, connection):
 		with connection.cursor() as cursor:
@@ -108,7 +71,7 @@ class database(object):
 				cursor.execute("SELECT * FROM services")
 				rows = cursor.fetchall()
 			except Exception as e:
-				self.log_error(e)
+				self.logging.log_error(e)
 
 			if(len(rows) > 0):
 				for service in rows:
